More fine grained evaluation results for GDA
Fixes bug that incorrectly aggregated a pin's users statistics (bug was
that the denominator of total # of attempts was only for one user's
attempts vs. # of attempts from all users with pin).
Now, I display fn/fp rates for each user - which are really good (most
equal errors rates of 5%)